backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "asgiref>=3.8.1",
    "flask>=3.1.1",
    "flask-bcrypt>=1.0.1",
    "flask-cors>=6.0.1",
    "flask-jwt-extended>=4.7.1",
    "flask-pymongo>=3.0.1",
    "flask-restful>=0.3.10",
    "mangum>=0.19.0",
    "numpy>=2.3.0",
    "pandas>=2.3.0",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.7.0",
]
